Prepare working version of `package:tflite_native` for Dart 2.6
origin/master is currently broken: The tests on windows fail because the
checked-in `lib/src/blobs/libtensorflowlite_c-win64.dll` library does
not seem to have the necessary symbols.

This branch is based on fbabf0b9 which is a working version of the
`package:tflite_native`. Any changes necessary for this code to work
with D26 ffi will be included in this branch.